week-2/day-1/twin/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-ollama>=1.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
